+++ design-system/range/examples/constellation/range-controlled.tsx
@@ -1,0 +1,16 @@
+import React, { useState } from 'react';
+
+import Range from '../../src';
+
+const RangeControlledExample = () => {
+  const [value, setValue] = useState(50);
+
+  return (
+    <>
+      <Range step={1} value={value} onChange={(value) => setValue(value)} />
+      <p>The current value is: {value}</p>
+    </>
+  );
+};
+
+export default RangeControlledExample;