+++ design-system/range/examples/constellation/range-default.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import Range from '../../src';
+
+const RangeDefaultExample = () => {
+  return <Range step={1} min={1} max={100} />;
+};
+
+export default RangeDefaultExample;