+++ design-system/range/examples/constellation/range-disabled.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import Range from '../../src';
+
+const RangeDisabledExample = () => {
+  return <Range isDisabled step={1} min={1} max={100} />;
+};
+
+export default RangeDisabledExample;