+++ design-system/range/examples/constellation/range-uncontrolled.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import Range from '../../src';
+
+const RangeUncontrolledExample = () => {
+  return (
+    <Range step={1} onChange={(value) => console.log('new value', value)} />
+  );
+};
+
+export default RangeUncontrolledExample;