+++ design-system/tokens/src/tokens/atlassian-dark/color/border.tsx
@@ -4,23 +4,8 @@ const color: BorderColorTokenSchema = {
   color: {
     border: {
       focus: {
-        value: [
-          {
-            color: 'B300',
-            offset: { x: 0, y: 0 },
-            opacity: 1,
-            radius: 0,
-            spread: 4,
-          },
-          {
-            color: 'DN0',
-            offset: { x: 0, y: 0 },
-            opacity: 1,
-            radius: 0,
-            spread: 2,
-          },
-        ],
-        attributes: { group: 'shadow' },
+        value: 'B300',
+        attributes: { group: 'paint' },
       },
       neutral: {
         value: 'DN300A',